# business/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controllers.createTicketRequest:
    properties:
      contactId:
        type: integer
      queueId:
        type: integer
      status:
        type: string
      whatsappId:
        type: integer
    required:
    - contactId
    type: object
...
host: localhost:8082
info:
  contact:
    email: dev@watink.com.br
    name: Watink Dev
  description: API do backend Go — atendimento e automação WhatsApp.
  license:
    name: Proprietary
    url: https://watink.com.br
  termsOfService: https://watink.com.br/terms
  title: Watink Business API
  version: 2.0.0
paths:
...
  /tickets:
    get:
      parameters:
      - description: Filtro por status (open, pending, closed)
        in: query
        name: status
        type: string
      - description: Filtro por fila
        in: query
        name: queueId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar tickets
      tags:
      - tickets
    post:
      consumes:
      - application/json
      description: Abre um ticket para um contato. Se já existir ticket aberto/pendente
        do contato na mesma conexão, devolve o existente (200). Sem whatsappId, usa
        a conexão do usuário ou a única conectada; com 2+ conectadas responde 409
        code=CONNECTION_REQUIRED listando as opções.
      parameters:
      - description: contactId obrigatório
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.createTicketRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Iniciar conversa (criar ticket)
      tags:
      - tickets
  /tickets/{ticketId}:
    delete:
      description: Remove definitivamente a conversa e suas mensagens. Deals vinculados
        via ticketId são preservados (só desvinculados, nunca apagados) — o funil
        de vendas não depende do ticket existir.
      parameters:
      - description: ID do ticket
        in: path
        name: ticketId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Excluir ticket
      tags:
      - tickets
    get:
      parameters:
      - description: ID do ticket
        in: path
        name: ticketId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Detalhar ticket
      tags:
      - tickets
    put:
      consumes:
      - application/json
      parameters:
      - description: ID do ticket
        in: path
        name: ticketId
        required: true
        type: integer
      - description: Campos a atualizar
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar ticket
      tags:
      - tickets
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token. Formato: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"